Workarounds for some linting issues
timu-line.el:492:34: Warning: the function ‘elfeed-search--count-unread’ is not known to be defined.
timu-line.el:326:16: Warning: the function ‘evil-visual-state-p’ is not known to be defined.
timu-line.el:325:16: Warning: the function ‘evil-replace-state-p’ is not known to be defined.
timu-line.el:324:16: Warning: the function ‘evil-operator-state-p’ is not known to be defined.
timu-line.el:323:16: Warning: the function ‘evil-normal-state-p’ is not known to be defined.
timu-line.el:322:16: Warning: the function ‘evil-motion-state-p’ is not known to be defined.
timu-line.el:321:16: Warning: the function ‘evil-insert-state-p’ is not known to be defined.
timu-line.el:320:16: Warning: the function ‘evil-emacs-state-p’ is not known to be defined.
timu-line.el:276:43: Warning: the function ‘f-dirname’ is not known to be defined.
timu-line.el:276:31: Warning: the function ‘f-filename’ is not known to be defined.
timu-line.el:276:23: Warning: the function ‘f-join’ is not known to be defined.